more function to method migration and refactoring of utils
